version 3

+++ app/src/main/java/com/kappdev/reminderwallpaper/wallpapers_feature/domain/model/Progress.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.graphics.Color
 
 data class Progress(
     val goal: String,
-    val complete: Int,
+    val complete: Float,
     val background: Color,
     val textColor: Color,
     val chartColor: Color
